refactor(r533): extract MPVLibProxy to MpvLibProxy.kt
Shared by PlayerFileLoadedHandler, PlayerMpvInitializer, and
PlayerActivity — doesn't belong inside PlayerFileLoadedHandler.
